+++ src/app/search/[id]/page.tsx
@@ -1,0 +1,3 @@
+export default function searchDetail() {
+  return <div>상세페이지입니다.</div>;
+}
