+++ src/app/search/page.tsx
@@ -1,9 +1,7 @@
 "use client";
 
-import { SearchTrack } from "@/types/search"; // 타입스크립트
-import { searchTracks } from "@/lib/spotifyToken"; // 토큰
-import { useEffect, useState } from "react";
-
+import { SearchTrack } from "@/types/search";
+import { useSearchData } from "@/hooks/useSearchData";
 import Pagination from "@/components/Pagination";
 import Image from "next/image";
 
@@ -15,45 +13,7 @@ export default function SearchPage({
   const query = Array.isArray(searchParams?.query) ? searchParams.query[0] : searchParams?.query || "";
   const pageQuery = searchParams?.page ? Number(searchParams.page) : 1;
 
-  const [results, setResults] = useState<SearchTrack[]>([]);
-
-  const [totalPages, setTotalPages] = useState(1);
-  const [currentPage, setCurrentPage] = useState(1);
-  const limit = 20;
-  const pageRange = 10;
-
-  // 전체 데이터 수 페이지네이션 생성
-  const fetchTotalPages = async (): Promise<void> => {
-    try {
-      const response = await searchTracks(query, 0, limit);
-      const { total } = response.tracks;
-      setTotalPages(Math.ceil(total / limit));
-    } catch (error) {
-      console.error("전체 데이터 가져오기 실패:", error);
-    }
-  };
-
-  // 해당 페이지의 데이터만 가져오기
-  const fetchData = async (page: number): Promise<void> => {
-    const offset = (page - 1) * limit;
-    try {
-      const response = await searchTracks(query, offset, limit);
-      const { items } = response.tracks;
-      setResults(items);
-    } catch (error) {
-      console.error("데이터 가져오기 실패:", error);
-    }
-  };
-
-  // query 바뀌면 실행
-  useEffect(() => {
-    if (query) {
-      setResults([]);
-      setCurrentPage(pageQuery);
-      fetchTotalPages();
-      fetchData(pageQuery);
-    }
-  }, [query, pageQuery]);
+  const { results, totalPages, currentPage, setCurrentPage, fetchData } = useSearchData(query, pageQuery);
 
   // 페이지 이동 함수
   const movePage = (page: number) => {
@@ -79,7 +39,7 @@ export default function SearchPage({
       </ul>
 
       <div style={{ marginTop: "20px" }}>
-        <Pagination totalPages={totalPages} currentPage={currentPage} pageRange={pageRange} movePage={movePage} />
+        <Pagination totalPages={totalPages} currentPage={currentPage} pageRange={10} movePage={movePage} />
       </div>
     </div>
   );
